Restored service hardening limits on server.
git-svn-id: https://opencadc.googlecode.com/svn/trunk@1457 728ff76a-78ac-11de-a72b-d90af8dea425

+++ projects/cadcVOS/src/ca/nrc/cadc/vos/server/DatabaseNodePersistence.java
@@ -118,7 +118,7 @@ public abstract class DatabaseNodePersistence implements NodePersistence
     protected NodeDAO.NodeSchema nodeSchema;
     protected boolean markDeleted;
 
-    protected Integer maxChildLimit = new Integer(1000000);
+    protected Integer maxChildLimit = new Integer(1000);
 
     /**
      * Constructor. This uses the default behaviour of deleting rows from the
